refactor: Consolidate setup scripts from 7 to 2 essential scripts
- Replace setup.py with improved automated version (complete setup with models)
- Replace quick_setup.py with no-models version (fast setup without downloads)
- Delete redundant scripts: setup_improved.py, ultra_quick_setup.py, quick_start_no_models.py, install.py, install_ollama.py
- Clear naming: setup.py (full) vs quick_setup.py (no models)
- Eliminate user confusion from having too many similar setup options
